Implement Polygon contains Line, for convex polys only.
Also, add a little logic to the Polygon constructor to automatically
weed out redundant points that don't change the angle of the boundary.
This will help to simplify testing whether a line is on the boundary of
the polygon, because we don't need to worry about a line that runs along
multiple line segments of the polygon.