Update run_test_on_update.sh to work with fswatch.
fswatch is a cross-platform utility that provides a similar
functionality to inotifywait on linux.  The script now tests to see if
you've got fswatch installed, and uses it if so.  Otherwise it tests to
see if you've got inotifywait, and uses that if so.  If neither are
found, the script errors out.