Fix up equality semantics.
Earlier, I had defined the __eq__ magic methods to use a "nearly equals"
function for convenience.  This is nice when doing comparisons but it
leads to undesirable results when hashing items and putting them into
sets.  We really need to keep __eq__ to doing "strictly equal"
comparisons, and use separate logic for "nearly equals".  Unfortunately
this means no binary operators for "nearly equals", we have to use
function calls.