Make Polygon hashable by standardising the boundary start point.
Add a new property method to Polygon, 'points_standard'.  This returns a
linear ring of the polygon's boundary but with the start point
standardised to the lowest 'x' value, and if there is a tie, lowest 'y'
value wins.

This means that we can compare two Polygons for equality by comparing
their standardised boundary, and we can also sensibly hash Polygons for
inclusion in sets.