Add class names to hash tuples.
This will help to avoid collisions in Collections between geometries
with different types but equal tuple representations.  For example, a
MultiPoint of A and B versus a Line from A to B would both have a tuple
of (A, B) but should not yield the same hash.  The same goes for a
Polygon and a LineString with the same sequence of points, etc.