Add a bbox property method to Collection.
Hook up the generic Geometry.union method to the new (and not yet fully
baked) top-level union function.  It can correctly produce unions where
the geometries are fully distinct, or fully covered by other geometries
in the union, but doesn't have any logic yet for partial intersections.

Also add an alternative path in Point.distance, if we are on an older
Python that doesn't have math.dist in its standard library.